requirements.txt
flask
ollama
langchain-community
langchain-chroma
langchain-ollama
pymupdf